projects/firecrawl_scraper/requirements.txt
fastapi
uvicorn
python-dotenv
firecrawl-py
